moved ListNode into its own file(s)
